+++ StabilityMatrix.Avalonia/ViewModels/CheckpointsPageViewModel.cs
@@ -10,6 +10,7 @@ using CommunityToolkit.Mvvm.ComponentModel;
 using CommunityToolkit.Mvvm.Input;
 using FluentAvalonia.UI.Controls;
 using NLog;
+using StabilityMatrix.Avalonia.Services;
 using StabilityMatrix.Avalonia.ViewModels.Base;
 using StabilityMatrix.Avalonia.ViewModels.CheckpointManager;
 using StabilityMatrix.Avalonia.Views;
@@ -30,24 +31,34 @@ public partial class CheckpointsPageViewModel : PageViewModelBase
     private readonly ISettingsManager settingsManager;
     private readonly ModelFinder modelFinder;
     private readonly IDownloadService downloadService;
+    private readonly INotificationService notificationService;
 
     public override string Title => "Checkpoints";
 
-    public override IconSource IconSource => new SymbolIconSource
-        {Symbol = Symbol.Notebook, IsFilled = true};
+    public override IconSource IconSource =>
+        new SymbolIconSource { Symbol = Symbol.Notebook, IsFilled = true };
 
     // Toggle button for auto hashing new drag-and-dropped files for connected upgrade
-    [ObservableProperty] private bool isImportAsConnected;
-    [ObservableProperty] private bool isLoading;
-    [ObservableProperty] private bool isIndexing;
-    
+    [ObservableProperty]
+    private bool isImportAsConnected;
+
+    [ObservableProperty]
+    private bool isLoading;
+
+    [ObservableProperty]
+    private bool isIndexing;
+
+    [ObservableProperty]
+    private bool showConnectedModelImages;
+
     [ObservableProperty]
     private string searchFilter = string.Empty;
 
     partial void OnIsImportAsConnectedChanged(bool value)
     {
-        if (settingsManager.IsLibraryDirSet &&
-            value != settingsManager.Settings.IsImportAsConnected)
+        if (
+            settingsManager.IsLibraryDirSet && value != settingsManager.Settings.IsImportAsConnected
+        )
         {
             settingsManager.Transaction(s => s.IsImportAsConnected = value);
         }
@@ -57,20 +68,24 @@ public partial class CheckpointsPageViewModel : PageViewModelBase
     private ObservableCollection<CheckpointManager.CheckpointFolder> checkpointFolders = new();
 
     [ObservableProperty]
-    private ObservableCollection<CheckpointManager.CheckpointFolder> displayedCheckpointFolders = new();
+    private ObservableCollection<CheckpointManager.CheckpointFolder> displayedCheckpointFolders =
+        new();
 
     public CheckpointsPageViewModel(
         ISharedFolders sharedFolders,
         ISettingsManager settingsManager,
         IDownloadService downloadService,
-        ModelFinder modelFinder)
+        INotificationService notificationService,
+        ModelFinder modelFinder
+    )
     {
         this.sharedFolders = sharedFolders;
         this.settingsManager = settingsManager;
         this.downloadService = downloadService;
+        this.notificationService = notificationService;
         this.modelFinder = modelFinder;
     }
-    
+
     public override async Task OnLoadedAsync()
     {
         var sw = Stopwatch.StartNew();
@@ -78,19 +93,23 @@ public partial class CheckpointsPageViewModel : PageViewModelBase
 
         // Set UI states
         IsImportAsConnected = settingsManager.Settings.IsImportAsConnected;
+        ShowConnectedModelImages = settingsManager.Settings.ShowConnectedModelImages;
         // Refresh search filter
         OnSearchFilterChanged(string.Empty);
 
-        Logger.Info($"Loaded {DisplayedCheckpointFolders.Count} checkpoint folders in {sw.ElapsedMilliseconds}ms");
-        
-        if (Design.IsDesignMode) return;
+        Logger.Info(
+            $"Loaded {DisplayedCheckpointFolders.Count} checkpoint folders in {sw.ElapsedMilliseconds}ms"
+        );
+
+        if (Design.IsDesignMode)
+            return;
 
         IsLoading = CheckpointFolders.Count == 0;
         IsIndexing = CheckpointFolders.Count > 0;
         await IndexFolders();
         IsLoading = false;
         IsIndexing = false;
-        
+
         Logger.Info($"OnLoadedAsync in {sw.ElapsedMilliseconds}ms");
     }
 
@@ -105,16 +124,16 @@ public partial class CheckpointsPageViewModel : PageViewModelBase
                 {
                     x.SearchFilter = SearchFilter;
                     return x;
-                }));
+                })
+            );
             sw.Stop();
             Logger.Info($"OnSearchFilterChanged in {sw.ElapsedMilliseconds}ms");
             return;
         }
-        
+
         sw.Restart();
-        
-        var filteredFolders = CheckpointFolders
-            .Where(ContainsSearchFilter).ToList();
+
+        var filteredFolders = CheckpointFolders.Where(ContainsSearchFilter).ToList();
         foreach (var folder in filteredFolders)
         {
             folder.SearchFilter = SearchFilter;
@@ -124,17 +143,30 @@ public partial class CheckpointsPageViewModel : PageViewModelBase
 
         DisplayedCheckpointFolders = new ObservableCollection<CheckpointFolder>(filteredFolders);
     }
-    
+
+    partial void OnShowConnectedModelImagesChanged(bool value)
+    {
+        if (
+            settingsManager.IsLibraryDirSet
+            && value != settingsManager.Settings.ShowConnectedModelImages
+        )
+        {
+            settingsManager.Transaction(s => s.ShowConnectedModelImages = value);
+        }
+    }
+
     private bool ContainsSearchFilter(CheckpointManager.CheckpointFolder folder)
     {
         if (folder == null)
             throw new ArgumentNullException(nameof(folder));
 
         // Check files in the current folder
-        return folder.CheckpointFiles.Any(x =>
-                   x.FileName.Contains(SearchFilter, StringComparison.OrdinalIgnoreCase)) ||
-               // If no matching files were found in the current folder, check in all subfolders
-               folder.SubFolders.Any(subFolder => ContainsSearchFilter(subFolder));
+        return folder.CheckpointFiles.Any(
+                x => x.FileName.Contains(SearchFilter, StringComparison.OrdinalIgnoreCase)
+            )
+            ||
+            // If no matching files were found in the current folder, check in all subfolders
+            folder.SubFolders.Any(subFolder => ContainsSearchFilter(subFolder));
     }
 
     private async Task IndexFolders()
@@ -146,48 +178,53 @@ public partial class CheckpointsPageViewModel : PageViewModelBase
             CheckpointFolders.Clear();
             return;
         }
-        
+
         // Setup shared folders in case they're missing
         sharedFolders.SetupSharedModelFolders();
 
         var folders = Directory.GetDirectories(modelsDirectory);
 
         var sw = Stopwatch.StartNew();
-        
+
         // Index all folders
-        var indexTasks = folders.Select(async f =>
-        {
-            var checkpointFolder =
-                new CheckpointFolder(settingsManager, downloadService, modelFinder)
+        var indexTasks = folders
+            .Select(async f =>
+            {
+                var checkpointFolder = new CheckpointFolder(
+                    settingsManager,
+                    downloadService,
+                    modelFinder,
+                    notificationService
+                )
                 {
                     Title = Path.GetFileName(f),
                     DirectoryPath = f,
                     IsExpanded = true, // Top level folders expanded by default
                 };
-            await checkpointFolder.IndexAsync();
-            return checkpointFolder;
-        }).ToList();
+                await checkpointFolder.IndexAsync();
+                return checkpointFolder;
+            })
+            .ToList();
 
         await Task.WhenAll(indexTasks);
-        
+
         sw.Stop();
         Logger.Info($"IndexFolders in {sw.ElapsedMilliseconds}ms");
 
         // Set new observable collection, ordered by alphabetical order
-        CheckpointFolders =
-            new ObservableCollection<CheckpointFolder>(indexTasks
-                .Select(t => t.Result)
-                .OrderBy(f => f.Title));
-        
+        CheckpointFolders = new ObservableCollection<CheckpointFolder>(
+            indexTasks.Select(t => t.Result).OrderBy(f => f.Title)
+        );
+
         if (!string.IsNullOrWhiteSpace(SearchFilter))
         {
             var filtered = CheckpointFolders
-                .Where(x => x.CheckpointFiles.Any(y => y.FileName.Contains(SearchFilter))).Select(
-                    f =>
-                    {
-                        f.SearchFilter = SearchFilter;
-                        return f;
-                    });
+                .Where(x => x.CheckpointFiles.Any(y => y.FileName.Contains(SearchFilter)))
+                .Select(f =>
+                {
+                    f.SearchFilter = SearchFilter;
+                    return f;
+                });
             DisplayedCheckpointFolders = new ObservableCollection<CheckpointFolder>(filtered);
         }
         else