+++ StabilityMatrix.Core/Models/Settings/Settings.cs
@@ -8,9 +8,9 @@ public class Settings
     public bool FirstLaunchSetupComplete { get; set; }
     public string? Theme { get; set; } = "Dark";
     public string? Language { get; set; } = "en-US";
-    
+
     public List<InstalledPackage> InstalledPackages { get; set; } = new();
-    
+
     [JsonPropertyName("ActiveInstalledPackage")]
     public Guid? ActiveInstalledPackageId { get; set; }
 
@@ -21,33 +21,34 @@ public class Settings
     [JsonIgnore]
     public InstalledPackage? ActiveInstalledPackage
     {
-        get => ActiveInstalledPackageId == null ? null : 
-            InstalledPackages.FirstOrDefault(x => x.Id == ActiveInstalledPackageId);
+        get =>
+            ActiveInstalledPackageId == null
+                ? null
+                : InstalledPackages.FirstOrDefault(x => x.Id == ActiveInstalledPackageId);
         set => ActiveInstalledPackageId = value?.Id;
     }
-    
+
     public bool HasSeenWelcomeNotification { get; set; }
     public List<string>? PathExtensions { get; set; }
     public string? WebApiHost { get; set; }
     public string? WebApiPort { get; set; }
-    
+
     // UI states
     public bool ModelBrowserNsfwEnabled { get; set; }
     public bool IsNavExpanded { get; set; }
     public bool IsImportAsConnected { get; set; }
+    public bool ShowConnectedModelImages { get; set; }
     public SharedFolderType? SharedFolderVisibleCategories { get; set; } =
-        SharedFolderType.StableDiffusion | 
-        SharedFolderType.Lora | 
-        SharedFolderType.LyCORIS;
+        SharedFolderType.StableDiffusion | SharedFolderType.Lora | SharedFolderType.LyCORIS;
 
     public WindowSettings? WindowSettings { get; set; }
 
     public ModelSearchOptions? ModelSearchOptions { get; set; }
-    
+
     public bool RemoveFolderLinksOnShutdown { get; set; }
-    
+
     public bool IsDiscordRichPresenceEnabled { get; set; }
-    
+
     public Dictionary<string, string>? EnvironmentVariables { get; set; }
 
     public HashSet<string>? InstalledModelHashes { get; set; } = new();
@@ -58,13 +59,13 @@ public class Settings
     {
         RemoveInstalledPackageAndUpdateActive(package.Id);
     }
-    
+
     public void RemoveInstalledPackageAndUpdateActive(Guid id)
     {
         InstalledPackages.RemoveAll(x => x.Id == id);
         UpdateActiveInstalledPackage();
     }
-    
+
     /// <summary>
     /// Update ActiveInstalledPackage if not valid
     /// uses first package or null if no packages